+++ frontend/scripts/forms.js
@@ -1,0 +1,5 @@
+Pragion.forms = {
+  validate(form) {
+    return form.checkValidity();
+  },
+};