+++ src/pragion/__init__.py
@@ -2,6 +2,6 @@
 
 from .core.app import App
 
-__version__ = "0.0.4"
+__version__ = "0.0.5"
 
 __all__ = ["App", "__version__"]